+++ src/router/index.js
@@ -3,12 +3,54 @@ import Router from 'vue-router'
 import store from '@/store/index'
 import PageRender from '@/components/PageRender'
 import LayoutAdmin from '@/components/admin/LayoutAdmin'
+import SignIn from '@/components/admin/login/SignIn'
+import SignUp from '@/components/admin/login/SignUp'
+import AdminHeader from '@/components/admin/ui/AdminHeader'
+import SideMenu from '@/components/admin/ui/SideMenu'
+import DashboardAdmin from '@/components/admin/DashboardAdmin'
+import ListPages from '@/components/admin/ListPages'
 
 Vue.use(Router)
 
 const router = new Router({
   mode: 'history',
   routes: [
+    // Sign In
+    {
+      path: '/login',
+      name: 'login',
+      component: LayoutAdmin,
+      props: {
+        header: false,
+        sidebar: false,
+        footer: false,
+      },
+      meta: {
+        requiresAuth: false,
+      },
+      children: [
+        {
+          path: 'sign-in',
+          name: 'signIn',
+          components: {
+            header: AdminHeader,
+            main: SignIn,
+          },
+          meta: {
+            requiresAuth: false,
+          },
+        },
+        {
+          path: 'sign-up',
+          name: 'signUp',
+          component: SignUp,
+          meta: {
+            header: AdminHeader,
+            requiresAuth: false,
+          },
+        },
+      ],
+    },
     // Admin
     {
       path: '/admin',
@@ -16,6 +58,51 @@ const router = new Router({
       meta: {
         requiresAuth: true,
       },
+      children: [
+        {
+          path: '',
+          name: 'AdminDashnoard',
+          components: {
+            header: AdminHeader,
+            main: DashboardAdmin,
+            sidebar: SideMenu,
+          },
+          meta: {
+            requiresAuth: false,
+          },
+        },
+        {
+          path: 'pages',
+          name: 'PagesList',
+          components: {
+            header: AdminHeader,
+            main: ListPages,
+            sidebar: SideMenu,
+          },
+          meta: {
+            requiresAuth: false,
+          },
+          children: [
+            {
+              path: 'add',
+              name: 'PagesAdd',
+              // component: InfluencersForm,
+              meta: {
+                requiresAuth: true,
+              },
+            },
+            {
+              path: 'edit/:slug',
+              name: 'PagesEdit',
+              // component: CategoriesForm,
+              props: true,
+              meta: {
+                requiresAuth: true,
+              },
+            },
+          ],
+        },
+      ],
     },
     // Public site
     {
@@ -35,7 +122,7 @@ router.beforeEach((to, from, next) => {
   const requiresAuth = to.matched.some(record => record.meta.requiresAuth)
 
   if (requiresAuth && !isLoggedIn) {
-    // go to login page
+    next('/login/sign-in')
   } else if (requiresAuth && isLoggedIn) {
     next()
   } else {